Acrescentado avisos de segurança ao tentar deletar rpg

+++ src/app/rpg/rpg-control/rpg-control.component.ts
@@ -111,19 +111,29 @@ export class RpgControlComponent implements OnInit {
 
   deleteRpg() {
     if (this.validateService.token() && this.validateService.master()) {
-      this.helperService.showLoading();
-      this.rpgService.delete(this.rpg.id)
-      .subscribe(
-        (response: HttpSuccessResponse) => {
-          this.helperService.showResponse(response);
-          this.rpgService.rpg(this.rpgId);
-          this.helperService.hideLoading();
-        },
-        (error: HttpErrorResponse) => {
-          console.log(error);
-          this.helperService.hideLoading();
+      this.helperService.openConfirm('Tem certeza que deseja apagar este rpg?')
+      .subscribe((result) => {
+        if (result) {
+          this.helperService.openConfirm('REALMENTE tem certeza que deseja apagar este rpg?!')
+          .subscribe((result) => { 
+            if (result) {
+              this.helperService.showLoading();
+              this.rpgService.delete(this.rpg.id)
+              .subscribe(
+                (response: HttpSuccessResponse) => {
+                  this.helperService.showResponse(response);
+                  this.rpgService.rpg(this.rpgId);
+                  this.helperService.hideLoading();
+                },
+                (error: HttpErrorResponse) => {
+                  console.log(error);
+                  this.helperService.hideLoading();
+                }
+              );
+            }
+          });
         }
-      );
+      });  
     }
   }
